fix: duplicate button styling
