+++ apps/web/components/Feature/ExploreCreators/Hero/styles.ts
@@ -1,5 +1,6 @@
 import { media } from "@repo/ui/breakpoints";
 import styled from "styled-components";
+import GenericButton from "@/components/UI/GenericButton";
 import { MonoText } from "@/components/UI/Monotext";
 
 export const Hero = styled.section`
@@ -149,7 +150,9 @@ export const FilterSection = styled.section`
   border-top: 1px solid ${({ theme }) => theme.colors.neutral.GRAY_200};
 `;
 
-export const FilterSectionButton = styled.button`
+export const FilterSectionButton = styled(GenericButton).attrs({
+  size: "md",
+})`
   width: 100%;
   padding: 18px 0;
   display: flex;
@@ -260,7 +263,9 @@ export const CheckboxControl = styled.span<{
   }
 `;
 
-export const ShowMoreButton = styled.button`
+export const ShowMoreButton = styled(GenericButton).attrs({
+  size: "sm",
+})`
   margin-top: 14px;
   padding: 0;
   border: none;
