+++ apps/web/components/Layout/Sidebar/styles.ts
@@ -2,6 +2,7 @@ import { MonoText } from "@/components/UI/Monotext";
 import { CREATOR_VARIANTS, CreatorVariant } from "@/utils/SidebarItems";
 import { media } from "@kiibee/ui/breakpoints";
 import styled from "styled-components";
+import GenericButton from "@/components/UI/GenericButton";
 
 export const SidebarWrapper = styled.aside<{ $open: boolean }>`
   width: 250px;
@@ -100,7 +101,9 @@ export const SidebarItemStyled = styled.div<{
   }
 `;
 
-export const MobileToggle = styled.button`
+export const MobileToggle = styled(GenericButton).attrs({
+  size: "sm",
+})`
   display: none;
   position: fixed;
   top: 35px;
