+++ apps/web/components/UI/GenericButton/styles.ts
@@ -1,7 +1,7 @@
-import styled, { css } from "styled-components";
+import styled from "styled-components";
 import type { Variant } from "@/utils/Constants";
-
-export type ButtonSize = "sm" | "md" | "lg";
+import { shared } from "./variants";
+import type { ButtonSize } from "./variants";
 
 type ButtonStyleProps = {
   $variant: Variant;
@@ -9,92 +9,6 @@ type ButtonStyleProps = {
   $fullWidth: boolean;
   $minWidth?: string;
 };
-
-const sizeStyles = {
-  sm: css`
-    min-height: 23px;
-    padding: 6px 14px;
-    ${({ theme }) => theme.typography.Body_Bold}
-  `,
-  md: css`
-    min-height: 40px;
-    padding: 7px 18px;
-    ${({ theme }) => theme.typography.Body_Medium}
-  `,
-  lg: css`
-    min-height: 49px;
-    padding: 14px 24px;
-    ${({ theme }) => theme.typography.Body_Medium}
-  `,
-};
-
-export const shared = css<ButtonStyleProps>`
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.5rem;
-  width: ${({ $fullWidth }) => ($fullWidth ? "100%" : "auto")};
-  min-width: ${({ $minWidth }) => $minWidth ?? "0"};
-  border-radius: 8px;
-  cursor: pointer;
-  text-decoration: none;
-  transition: all 120ms ease;
-  ${({ $size }) => sizeStyles[$size]}
-
-  &:disabled,
-  &[aria-disabled="true"] {
-    cursor: not-allowed;
-    opacity: 0.55;
-    pointer-events: none;
-  }
-
-  ${({ $variant }) => {
-    switch ($variant) {
-      case "primary":
-        return css`
-          background: ${({ theme }) => theme.colors.primary.BLACK};
-          color: ${({ theme }) => theme.colors.primary.WHITE};
-          border: 1px solid ${({ theme }) => theme.colors.primary.BLACK};
-          box-shadow: 0 6px 18px rgba(0, 0, 0, 0.2);
-          &:hover {
-            background: transparent;
-            color: ${({ theme }) => theme.colors.primary.BLACK};
-            box-shadow: none;
-            transform: none;
-            opacity: 1;
-          }
-        `;
-
-      case "primary-lite":
-        return css`
-          background: ${({ theme }) => theme.colors.secondary.MEDIUM_GREEN};
-          color: ${({ theme }) => theme.colors.primary.BLACK};
-          border: 1px solid transparent;
-          border-radius: 0.5rem;
-          &:hover {
-            background: transparent;
-            border: 1px solid ${({ theme }) => theme.colors.primary.BLACK};
-            opacity: 1;
-          }
-        `;
-
-      default:
-        return css`
-          background: transparent;
-          color: ${({ theme }) => theme.colors.primary.BLACK};
-          border: 1px solid ${({ theme }) => theme.colors.primary.BLACK};
-          &:hover {
-            background: ${({ theme }) => theme.colors.primary.BLACK};
-            color: ${({ theme }) => theme.colors.primary.WHITE};
-            border: 1px solid ${({ theme }) => theme.colors.primary.BLACK};
-            box-shadow: 0 6px 18px rgba(0, 0, 0, 0.2);
-            transform: translateY(-1px);
-            opacity: 0.98;
-          }
-        `;
-    }
-  }}
-`;
 
 export const ButtonEl = styled.button<ButtonStyleProps>`
   ${shared}
