+++ apps/web/components/UI/SearchBar/styles.ts
@@ -1,5 +1,6 @@
 import { media } from "@repo/ui/breakpoints";
 import styled from "styled-components";
+import GenericButton from "@/components/UI/GenericButton";
 
 export const Wrapper = styled.div<{ $width?: string }>`
   width: ${({ $width }) => $width || "100%"};
@@ -43,7 +44,9 @@ export const StyledInput = styled.input`
   }
 `;
 
-export const ClearButton = styled.button`
+export const ClearButton = styled(GenericButton).attrs({
+  size: "sm",
+})`
   display: inline-flex;
   align-items: center;
   justify-content: center;
